requirements.txt
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
numpy==2.2.6
opencv_python==4.12.0.88
openpyxl==3.1.5
pandas==2.3.3
pdf2image==1.17.0
Pillow==12.0.0
PyPDF2==3.0.1
PyYAML==6.0.3
Requests==2.32.5
streamlit==1.51.0
streamlit_calendar==1.4.0
streamlit_extras==0.7.8
streamlit_modal==0.1.2
streamlit_option_menu==0.4.0
azure-ai-documentintelligence==1.0.2
azure-core==1.36.0 
psycopg2==2.9.11
psycopg2-binary==2.9.11